refactor calculator to another file
