+++ calculator.js
@@ -1,0 +1,16 @@
+const calculator = {
+  add: (a, b) => {
+    return a + b;
+  },
+  subtract: (a, b) => {
+    return a - b;
+  },
+  divide: (a, b) => {
+    return a / b;
+  },
+  multiply: (a, b) => {
+    return a * b;
+  },
+};
+
+export { calculator };
